pass template state to form component to open dialog properly
- #330
