+++ frontend/src/features/rubricBuilder/TemplateUpload.tsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useState } from "react";
-import { Template } from "palette-types";
+import { Criteria, Template } from "palette-types";
 import { useFetch } from "@hooks";
+import { useTemplate } from "../../hooks/useTemplate.ts";
+import { useChoiceDialog, useRubric } from "@context";
 
 interface TemplateUploadProps {
   closeImportCard: () => void; // callback to close the template import card
-  onTemplateSelected: (template: Template) => void;
 }
 
 const TemplateUpload: React.FC<TemplateUploadProps> = ({
   closeImportCard,
-  onTemplateSelected,
 }: TemplateUploadProps) => {
   const [templates, setTemplates] = useState<Template[]>([]);
 
@@ -44,6 +44,73 @@ const TemplateUpload: React.FC<TemplateUploadProps> = ({
     }
 
     closeImportCard();
+  };
+
+  const { setUpdatingTemplate, setImportingTemplate } = useTemplate();
+  const { openDialog } = useChoiceDialog();
+  const { activeRubric, setActiveRubric } = useRubric();
+
+  const onTemplateSelected = (template: Template) => {
+    const updatedTemplate = {
+      ...template,
+      usageCount: template.usageCount + 1,
+      lastUsed: new Date().toISOString(),
+    };
+
+    setUpdatingTemplate(updatedTemplate);
+    if (!activeRubric) return;
+
+    const currentCriteria = activeRubric.criteria;
+    const newCriteria = template.criteria;
+
+    if (newCriteria.length === 0) {
+      openDialog({
+        title: "No Criteria Detected",
+        message: "This template has no criteria",
+        buttons: [],
+      });
+      return;
+    }
+
+    // Split into unique and duplicate criteria
+    const { unique, duplicates } = newCriteria.reduce(
+      (acc, newCriterion) => {
+        const isDuplicate = currentCriteria.some(
+          (existingCriterion) =>
+            existingCriterion.key.trim().toLowerCase() ===
+            newCriterion.key.trim().toLowerCase(),
+        );
+
+        if (isDuplicate) {
+          acc.duplicates.push(newCriterion);
+        } else {
+          acc.unique.push(newCriterion);
+        }
+
+        return acc;
+      },
+      { unique: [] as Criteria[], duplicates: [] as Criteria[] },
+    );
+
+    // Log information about duplicates if any were found
+    if (duplicates.length > 0) {
+      const duplicateDescriptions = duplicates
+        .map((criterion) => criterion.description)
+        .join(", ");
+
+      openDialog({
+        title: "Duplicate Criteria Detected",
+        message: `Looks like you already imported this one. Duplicate criteria: ${duplicateDescriptions}`,
+        buttons: [],
+      });
+      return;
+    }
+
+    setActiveRubric({
+      ...activeRubric,
+      criteria: [...currentCriteria, ...unique],
+    });
+    setImportingTemplate(updatedTemplate);
   };
 
   return (
